+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Removed follower count from profile screen.
+- Decode nprofile, nevent, and naddr NIP-19 entities.
 
 ## [0.1.20] - 2024-07-10Z